backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
fpdf
pytz
pydantic[email]
reportlab
openpyxl
bcrypt==4.0.1
icalendar
python-dateutil
holidays>=0.35
